fix(windows): cast *mut c_void to isize for windows-sys HWND compatibility
Tauri's `windows` crate defines `HWND` as `pub struct HWND(pub *mut core::ffi::c_void)`,
so `h.0` yields a raw pointer — not `isize`. `windows-sys 0.59` defines
`type HWND = isize`, so a `h.0 as isize` cast is required at the boundary.

The previous refactor commit corrected the trait return type to `isize` but
missed the cast, causing a type-mismatch compile error on Windows CI.

Also corrects the doc comment that incorrectly described the `windows` crate
`HWND` as holding an `isize` field.

+++ src-tauri/src/lib.rs
@@ -1713,9 +1713,8 @@ fn copy_sample_files(handle: &tauri::AppHandle) -> Result<(), String> {
 /// Abstracts over the Tauri window types that expose a native `hwnd()` on
 /// Windows. `setup()` hands us a `WebviewWindow`; the `on_window_event` handler
 /// hands us a `Window`. Both expose `hwnd()` returning a `windows`-crate `HWND`
-/// (a `pub struct HWND(pub isize)` newtype); `.0` extracts the raw `isize`, which
-/// is the same representation as `windows-sys 0.59`'s `type HWND = isize`, so no
-/// intermediate pointer hop is needed.
+/// (`pub struct HWND(pub *mut core::ffi::c_void)`). We cast `.0 as isize` to
+/// match `windows-sys 0.59`'s `type HWND = isize`.
 #[cfg(target_os = "windows")]
 trait RawHwnd {
     fn raw_hwnd(&self) -> Result<isize, String>;
@@ -1724,14 +1723,14 @@ trait RawHwnd {
 #[cfg(target_os = "windows")]
 impl RawHwnd for tauri::WebviewWindow {
     fn raw_hwnd(&self) -> Result<isize, String> {
-        self.hwnd().map(|h| h.0).map_err(|e| e.to_string())
+        self.hwnd().map(|h| h.0 as isize).map_err(|e| e.to_string())
     }
 }
 
 #[cfg(target_os = "windows")]
 impl RawHwnd for tauri::Window {
     fn raw_hwnd(&self) -> Result<isize, String> {
-        self.hwnd().map(|h| h.0).map_err(|e| e.to_string())
+        self.hwnd().map(|h| h.0 as isize).map_err(|e| e.to_string())
     }
 }
